move implementation of ebpf_domain_t to ebpf_domain.cpp
