+++ src/crab/fwd_analyzer.hpp
@@ -11,7 +11,6 @@
 
 namespace crab {
 
-using domains::ebpf_domain_t;
 using invariant_table_t = std::map<label_t, ebpf_domain_t>;
 
 std::pair<invariant_table_t, invariant_table_t> run_forward_analyzer(cfg_t& cfg, const ebpf_domain_t& entry_inv,
